pinecone-data-lambda/requirements.txt
drive==0.4.4
google-api-python-client
langchain-openai
python-dotenv
langchain_pinecone
langchain-text-splitters
langchain_community
pinecone
